requirements.txt
python-binance
python-dotenv
python-settings
asyncio